add security to API - add extension method to ClaimsPrincipal for the quest by scope - add authorization requirements needed - add authorization handlers for the requirements - create representations for the identifiers of the secured resources - add Authorization Policies based on Requirements - use Policies on access of the resources

+++ Server/API/Controllers/ImagesController.cs
@@ -1,7 +1,10 @@
-﻿using API.Filters;
+﻿using Authorization;
+using Authorization.Resources;
+using API.Filters;
 using API.Interfaces.IServices;
 using IODomain.Input;
 using IODomain.Output;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -11,37 +14,61 @@ namespace API.Controllers {
     [Route("api/boards/{boardId}/figures/[controller]")]
     public class ImagesController : Controller {
         private readonly IImageService _imageService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ImagesController(IImageService imageService) {
+        public ImagesController(IImageService imageService, IAuthorizationService authorizationService) {
             _imageService = imageService;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
-        public Task<IEnumerable<OutImage>> GetAll(long boardId) {
-            return _imageService.GetAllAsync(boardId);
+        public async Task<IActionResult> GetAll(long boardId) {
+            if(!await _authorizationService.AuthorizeAsync(User, new BoardRequest(boardId), Policies.ReadBoardPolicy))
+                return new ChallengeResult();
+
+            IEnumerable < OutImage > images = await _imageService.GetAllAsync(boardId);
+
+            return Ok(images);
         }
 
         [HttpGet("{id}", Name = "GetImage")]
-        public Task<OutImage> GetById(long id, long boardId) {
-            return _imageService.GetAsync(id, boardId);
+        public async Task<IActionResult> GetById(long id, long boardId) {
+            if(!await _authorizationService.AuthorizeAsync(User, new BoardRequest(boardId), Policies.ReadBoardPolicy))
+                return new ChallengeResult();
+
+            OutImage image = await _imageService.GetAsync(id, boardId);
+
+            return Ok(image);
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(long boardId, [FromBody] InImage inputImage) {
+            if(!await _authorizationService.AuthorizeAsync(User, new BoardRequest(boardId), Policies.WriteBoadPolicy))
+                return new ChallengeResult();
+
             OutImage image = await _imageService.CreateAsync(boardId, inputImage);
+
             return CreatedAtRoute("GetImage", new { id = image.Id, boardId = boardId }, image);
         }
 
 
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(long id, long boardId, [FromBody] InImage inputImage) {
+            if(!await _authorizationService.AuthorizeAsync(User, new BoardRequest(boardId), Policies.WriteBoadPolicy))
+                return new ChallengeResult();
+
             await _imageService.UpdateAsync(id, boardId, inputImage);
+
             return new NoContentResult();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(long id, long boardId) {
+            if(!await _authorizationService.AuthorizeAsync(User, new BoardRequest(boardId), Policies.WriteBoadPolicy))
+                return new ChallengeResult();
+
             await _imageService.DeleteAsync(id, boardId);
+
             return new NoContentResult();
         }
     }
