+++ Server/API/Controllers/UsersController.cs
@@ -1,7 +1,10 @@
-﻿using API.Filters;
+﻿using Authorization;
+using Authorization.Resources;
+using API.Filters;
 using API.Interfaces.IServices;
 using IODomain.Input;
 using IODomain.Output;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -12,10 +15,12 @@ namespace API.Controllers {
     public class UsersController : Controller {
         private readonly IUserService _userService;
         private readonly IUsersBoardsService _usersBoardsService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public UsersController(IUserService userService, IUsersBoardsService usersBoardsService) {
+        public UsersController(IUserService userService, IUsersBoardsService usersBoardsService, IAuthorizationService authorizationService) {
             _userService = userService;
             _usersBoardsService = usersBoardsService;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -36,24 +41,40 @@ namespace API.Controllers {
 
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(long id, [FromBody] InUser inputUser) {
+            if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(""), Policies.UserIsOwnPolicy)) // REVER UserId parametro
+                return new ChallengeResult();
+
             await _userService.UpdateAsync(id, inputUser);
             return new NoContentResult();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(long id) {
+            if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(""), Policies.UserIsOwnPolicy)) // REVER UserId parametro
+                return new ChallengeResult();
+
             await _userService.DeleteAsync(id);
             return new NoContentResult();
         }
 
         [HttpGet("{userId}/boards")]
-        public Task<IEnumerable<OutUserBoard_Board>> GetAll(long userId, string search, long index = 0, long size = 10) {
-            return _usersBoardsService.GetAllBoardsAsync(userId, index, size, search);
+        public async Task<IActionResult> GetAll(long userId, string search, long index = 0, long size = 10) {
+            if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(""), Policies.UserIsOwnPolicy)) // REVER UserId parametro
+                return new ChallengeResult();
+
+            IEnumerable<OutUserBoard_Board> userBoards = await _usersBoardsService.GetAllBoardsAsync(userId, index, size, search);
+
+            return Ok(userBoards);
         }
 
         [HttpGet("{userId}/boards/{boardId}")]
-        public Task<OutUserBoard_Board> GetById(long userId, long boardId) {
-            return _usersBoardsService.GetBoardAsync(userId, boardId);
+        public async Task<IActionResult> GetById(long userId, long boardId) {
+            if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(""), Policies.UserIsOwnPolicy)) // REVER UserId parametro
+                return new ChallengeResult();
+
+            OutUserBoard_Board userBoard = await _usersBoardsService.GetBoardAsync(userId, boardId);
+
+            return Ok(userBoard);
         }
     }
 }
